introduce RmControl trait
Each control has a code, a parameter type, and a reponse type. By
associating them all using a trait, we just need to pass the parameter
type to `send_control`, and all the rest is derived from the trait. This
nicely models the control layer on top of the message one.