collapse RmApi into GspCmdq
RmApi is just a way to hold parameters, and is so constrained by its
lifetime parameter that we would need to repeatedly build it anyway.

Replace it by specialized impl blocks on GspCmdq for sending commands
and controls respectively. At the end of the day, this results in less
parameters being sent and highlights a few improvements we should have
in GspCmdq (like holding an ARef to the Device).