optional: use type aliases instead of newtypes for RM commands
This allows us to piggyback on the GspCommandElement implementation of
RmMessage instead of having to reimplement our own.

This comes with a caveat though: if we ever need to have different RM
commands that use the same header type, we will need a newtype because
that is the only way to provide a distinct implementation of RmCommand
for both. But so far that doesn't seem to be the case.