advance the queue's rptr on message drop
Having the explicitly ack the messages to advance the rptr sounds
redundant: there is no case for not processing a message we read, and
forgetting to ack messages has the potential to create infinite loops.

So just movethe rptr in the `GspQueueMessage` destructor. This way
`receive_msg` can be used in a while loop similarly to an iterator
(unfortunately, we cannot implement an iterator for the message queue
due to the lending iterator problem).