make cpu/gsp_rptr/wptr infallible
The only case for dma_read/write to fail is if we provide an invalid
index for the accessed object. Here we always access object 0, which we
know we have properly allocated, so there is no point in returning an
error that can never happen in these frequently-called methods.

Use `unwrap_unchecked` in an unsafe block to eliminate the panic code
path.