Add `DmaGspMem` type
The `GspMem` object is shared with the GSP and must be manipulated with
absolute care to avoid trespassing on Rust's memory safety guarantee.
Add a dedicated type to contain it that will provide safe methods to
provide access to the data that is legit to read or write by the driver.

We are still making unsafe accesses to the GspMem object in the code,
but these will be gradually removed as functionality is moved to
`DmaGspMem`.